qr code added . video added

+++ src/main/java/ua/shortener/link/service/LinkService.java
@@ -2,6 +2,7 @@ package ua.shortener.link.service;
 
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 
 import org.springframework.stereotype.Service;
 
@@ -17,8 +18,10 @@ import java.util.Optional;
 import java.util.Map;
 import java.util.HashMap;
 
+
 @Service
 @RequiredArgsConstructor
+@Slf4j
 public final class LinkService {
 
     private final LinkRepository linkRepository;
@@ -36,7 +39,13 @@ public final class LinkService {
     }
 
     public void deleteLink(final String shortLink) {
-        linkRepository.deleteById(shortLink);
+        if(!linkRepository.existsById(shortLink)){
+            throw new IllegalArgumentException("Link with id " + shortLink + " not found");
+        }else{
+            linkRepository.deleteById(shortLink);
+            log.info("Note by " + shortLink + " was deleted");
+        }
+
     }
 
     public Link editLink(final Link existingLink) {
